miles-platformer: initial commit
This is the initial commit for a simplistic 2d platformer, starring
none other than miles meraki. We chose to use the Quintus game engine,
primarily for simplicity.

This commit includes the following:
- HTML shim for loading the game
- Tiles, sprites, and assets for display
- Core game file with one level
- Quintus library at f66f40ab5485a9efc6a02c646f132633b01a9a62
  (so that we can include tmx level file support, for level creation
   via the Tiled tool)
- mkidata.js, which allows us to load sprite and level data from a file,
  rather than via XHR. Quintus defaults to XHR, so this is a workaround
  for running the game locally.
- gitignore file, so I don't commit vim .swp files.

The game includes Miles, a basic 'Enemy' class, a 'Qbox' class that functions
similarly to the question boxes in Super Mario, and a 'Powerup' class that
turns Miles invincible for 5 seconds (with animation!).